+++ src/main/java/com/smartstudy/service/UserService.java
@@ -1,53 +1,67 @@
 package com.smartstudy.service;
 
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
-import com.smartstudy.model.User;
+import com.smartstudy.model.MyUser;
 import com.smartstudy.repository.UserRepo;
 import com.smartstudy.utils.ModelUtils;
 import com.smartstudy.utils.Responce;
 
 
 @Service
-public class UserService {
+public class UserService implements UserDetailsService {
      @Autowired
      private UserRepo userRepo;
 
-     public Responce<User> save(User user){
+     public Responce<MyUser> save(MyUser user){
           try {
-               User savedUser = userRepo.getReferenceById(user.getEmail());
-               if(savedUser != null){
+               if(user == null) {
+                    return new Responce<MyUser>("Invalid user information");
+               }
+               Optional<MyUser> savedUser = userRepo.findById(user.getEmail());
+               if(savedUser.isPresent()){
                     return new Responce<>("User Already exists");
                }else{
-                    savedUser = userRepo.save(user);
-                    return new Responce<User>(savedUser, "User account created successfully");
+                    MyUser sUser = userRepo.save(user);
+                    return new Responce<MyUser>(sUser, "User account created successfully");
                }
           } catch (Exception e) {
                return new Responce<>(e,"Server Error");
           }
      }
 
-     public Responce<User> update(User user){
+     public Responce<MyUser> update(MyUser user){
           try {
-               User User = userRepo.getReferenceById(user.getEmail());
-               if(User == null){
+               if(user == null) {
+                    return new Responce<MyUser>("Invalid user information");
+               }
+               Optional<MyUser> ckUser = userRepo.findById(user.getEmail());
+               if(!ckUser.isPresent()){
                     return new Responce<>("User not found");
                }else{
-                    BeanUtils.copyProperties(user, User, ModelUtils.getNullPropertyNames(user));
-                    userRepo.save(User);
-                    return new Responce<User>(User, "user updated successfully");
+                    MyUser sUser = ckUser.get();
+                    BeanUtils.copyProperties(user, sUser,ModelUtils.getNullPropertyNames(user));
+                    userRepo.save(sUser);
+                    return new Responce<MyUser>(sUser, "user updated successfully");
                }
           } catch (Exception e) {
                return new Responce<>(e,"Server Error");
           }
      }
 
-     public Responce<User> delete(User user){
+     public Responce<MyUser> delete(MyUser user){
           try{
+               if(user == null) {
+                    return new Responce<MyUser>("Invalid user information");
+               }
                userRepo.delete(user);
                return new Responce<>("Deleted  the user successfully");
           }catch(Exception e){
@@ -55,13 +69,18 @@ public class UserService {
           }
      }
 
-     public Responce<User> search(User user){
+     public Responce<MyUser> search(MyUser user){
           try {
-               User foundUser = userRepo.findById(user.getEmail()).get();
-               if(foundUser == null){
-                    return new Responce<>("Cannot find the user searched");
-               }else {
+               if(user == null) {
+                    return new Responce<MyUser>("Invalid user information");
+               }
+               Optional<MyUser> oUser = userRepo.findById(user.getEmail());
+               if(oUser.isPresent()){
+                    MyUser foundUser = oUser.get();
                     return new Responce<>(foundUser, "user found");
+               }
+               else{
+                    return new Responce<>("Cannot find the user searched");
                }
           } catch (Exception e) {
                e.printStackTrace();
@@ -69,12 +88,22 @@ public class UserService {
           }
      }
 
-     public List<User> findAll(){
+     public List<MyUser> findAll(){
           try {
                return userRepo.findAll();
           } catch (Exception e) {
                e.printStackTrace();
                return null;
+          }
+     }
+
+     @Override
+     public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
+          Optional<MyUser> user = userRepo.findById(username);
+          if (user.isPresent()) {
+               return user.get();
+          } else {
+               throw new UsernameNotFoundException("User not found with username: " + username);
           }
      }
 }
